requirements.txt
certifi==2024.7.4
charset-normalizer==2.0.12
idna==3.7
pytz==2021.3
PyYAML==6.0.1
requests==2.26.0
urllib3==1.26.7
